bpf: Remove POLICY_MAP from bpf_netdev and bpf_overlay
Both the bpf_netdev and bpf_overlay have not been using a policy map for their
functionality but a BPF map was still allocated and pinned to the filesystem.
Refactor the code a bit to avoid allocating BPF policy maps for these programs.

Also, remove old pinned BPF maps on agent bootstrap.

Fixes: #3450
