+++ bpf/bpf_lxc.c
@@ -42,6 +42,7 @@
 #include "lib/l3.h"
 #include "lib/lxc.h"
 #include "lib/nat46.h"
+#include "lib/identity.h"
 #include "lib/policy.h"
 #include "lib/lb.h"
 #include "lib/drop.h"
