+++ bpf/bpf_netdev.c
@@ -42,9 +42,9 @@
 #include "lib/eth.h"
 #include "lib/dbg.h"
 #include "lib/trace.h"
+#include "lib/identity.h"
 #include "lib/l3.h"
 #include "lib/l4.h"
-#include "lib/policy.h"
 #include "lib/drop.h"
 #include "lib/encap.h"
 #include "lib/nat.h"
