+++ bpf/lib/policy.h
@@ -25,63 +25,6 @@
 #include "eps.h"
 #include "maps.h"
 
-static inline bool __inline__ inherit_identity_from_host(struct __sk_buff *skb, __u32 *identity)
-{
-	__u32 magic = skb->mark & MARK_MAGIC_HOST_MASK;
-	bool from_proxy = false;
-
-	/* Packets from the ingress proxy must skip the proxy when the
-	 * destination endpoint evaluates the policy. As the packet
-	 * would loop and/or the connection be reset otherwise. */
-	if (magic == MARK_MAGIC_PROXY_INGRESS) {
-		*identity = get_identity(skb);
-		skb->tc_index |= TC_INDEX_F_SKIP_INGRESS_PROXY;
-		from_proxy = true;
-	/* (Return) packets from the egress proxy must skip the
-	 * redirection to the proxy, as the packet would loop and/or
-	 * the connection be reset otherwise. */
-	} else if (magic == MARK_MAGIC_PROXY_EGRESS) {
-		*identity = get_identity(skb);
-		skb->tc_index |= TC_INDEX_F_SKIP_EGRESS_PROXY;
-		from_proxy = true;
-	} else if (magic == MARK_MAGIC_IDENTITY) {
-		*identity = get_identity(skb);
-	} else if (magic == MARK_MAGIC_HOST) {
-		*identity = HOST_ID;
-	} else {
-		*identity = WORLD_ID;
-	}
-
-	/* Reset packet mark to avoid hitting routing rules again */
-	skb->mark = 0;
-	cilium_dbg(skb, DBG_INHERIT_IDENTITY, *identity, 0);
-
-	return from_proxy;
-}
-
-
-/**
- * identity_is_reserved is used to determine whether an identity is one of the
- * reserved identities that are not handed out to endpoints.
- *
- * Specifically, it should return true if the identity is one of these:
- * - IdentityUnknown
- * - ReservedIdentityHost
- * - ReservedIdentityWorld
- * - ReservedIdentityRemoteNode
- *
- * The following identities are given to endpoints so return false for these:
- * - ReservedIdentityUnmanaged
- * - ReservedIdentityHealth
- * - ReservedIdentityInit
- *
- * Identities 128 and higher are guaranteed to be generated based on user input.
- */
-static inline bool identity_is_reserved(__u32 identity)
-{
-	return identity < UNMANAGED_ID || identity == REMOTE_NODE_ID;
-}
-
 #ifdef SOCKMAP
 static inline int __inline__
 policy_sk_egress(__u32 identity, __u32 ip,  __u16 dport)
