+++ pkg/datapath/maps/map.go
@@ -163,6 +163,15 @@ func (ms *MapSweeper) RemoveDisabledMaps() {
 			"cilium_proxy4"}...)
 	}
 
+	// Can be removed with Cilium 1.8
+	maps = append(maps, []string{
+		"cilium_policy_reserved_1",
+		"cilium_policy_reserved_2",
+		"cilium_policy_reserved_3",
+		"cilium_policy_reserved_4",
+		"cilium_policy_reserved_5",
+	}...)
+
 	for _, m := range maps {
 		p := path.Join(bpf.MapPrefixPath(), m)
 		if _, err := os.Stat(p); !os.IsNotExist(err) {
